More database consistency checks
* NEWS: Update.
* THANKS: Update.
* src/bucket.c (_gdbm_get_bucket): Check if directory entry is
valid. Don't cache invalid buckets.
* src/gdbm.h.in (GDBM_BAD_DIR_ENTRY): New error code.
* src/gdbmerrno.c: Likewise.
* src/gdbmopen.c (validate_header): Compute expected
number of bucket elements based on the bucket size, not on
the block size.
(_gdbm_init_cache_entry): New function.
* src/proto.h (_gdbm_init_cache_entry): New proto.
* src/recover.c (gdbm_recover): Clear error state after return
from check_db indicating failure.

+++ src/gdbmdefs.h
@@ -225,7 +225,7 @@ struct gdbm_file_info
   /* The bucket cache. */
   cache_elem *bucket_cache;
   size_t cache_size;
-  int last_read;
+  size_t last_read;
 
   /* Points to the current hash bucket in the cache. */
   hash_bucket *bucket;
